chore(ui): refactor metrics chart to generic-charts
