+++ components/generic-charts/card-metric.cy.tsx
@@ -8,10 +8,8 @@ describe('<CardMetric />', () => {
     cy.get('div').contains('0')
     cy.get('div').contains('10')
 
-    // Tremor bar should be 0% filled
-    cy.get('.tremor-ProgressBar-progressBar')
-      .should('have.attr', 'style')
-      .and('include', '0%')
+    // should be 0% filled
+    cy.get('span').should('have.attr', 'title').and('include', '0%')
   })
 
   it('renders 10% bar', () => {
@@ -21,10 +19,8 @@ describe('<CardMetric />', () => {
     cy.get('div').contains('1')
     cy.get('div').contains('10')
 
-    // Tremor bar should be 10% filled
-    cy.get('.tremor-ProgressBar-progressBar')
-      .should('have.attr', 'style')
-      .and('include', '10%')
+    // should be 10% filled
+    cy.get('span').should('have.attr', 'title').and('include', '10%')
   })
 
   it('renders 100% bar', () => {
@@ -34,10 +30,8 @@ describe('<CardMetric />', () => {
     cy.get('div').contains('10')
     cy.get('div').contains('10')
 
-    // Tremor bar should be 100% filled
-    cy.get('.tremor-ProgressBar-progressBar')
-      .should('have.attr', 'style')
-      .and('include', '100%')
+    // should be 100% filled
+    cy.get('span').should('have.attr', 'title').and('include', '100%')
   })
 
   it('renders 1000% bar when current > target', () => {
@@ -47,10 +41,8 @@ describe('<CardMetric />', () => {
     cy.get('div').contains('100')
     cy.get('div').contains('10')
 
-    // Tremor bar should be 1000% filled
-    cy.get('.tremor-ProgressBar-progressBar')
-      .should('have.attr', 'style')
-      .and('include', '1000%')
+    // should be 1000% filled
+    cy.get('span').should('have.attr', 'title').and('include', '1000%')
   })
 
   it('renders 10% bar with currentReadable and targetReadable', () => {
@@ -67,10 +59,8 @@ describe('<CardMetric />', () => {
     cy.get('div').contains('xx 1 xx')
     cy.get('div').contains('xx 10 xx')
 
-    // Tremor bar should be 10% filled
-    cy.get('.tremor-ProgressBar-progressBar')
-      .should('have.attr', 'style')
-      .and('include', '10%')
+    // should be 10% filled
+    cy.get('span').should('have.attr', 'title').and('include', '10%')
   })
 
   it('renders with custom className', () => {
@@ -82,7 +72,6 @@ describe('<CardMetric />', () => {
       />
     )
 
-    // Tremor bar should be 10% filled
     cy.get('.bg-cyan-400.custom-classname').should('exist')
   })
 })
